Keys to process stratigraphic, environmental and lithological information from the Paleobiology Database

Lists of entries treated as indicators of similar characteristics

Usage

data(keys)

Format

A list of 7 lists:

tenInt

A list of vectors. Entries in the early_interval and late_interval variables of PaleoDB downloads indicate the collections' positions in the dynamic time scale. These entries were linked to 10 million year-resolution time scale stored in tens. These links were compiled using a download from the FossilWorks website (used to be http://www.fossilworks.org/), on 08 June, 2018. You can check the lookup table stratkeys here. This is version 0.9.2

stgInt

A list of vectors. Entries in the early_interval and late_interval variables of PaleoDB downloads indicate the collections' positions in the dynamic time scale. These entries were linked to stage-resolution time scale stored in stages. See binInt for version information.

These entries are reliable only in the Post-Ordovician!
reefs

A list of vectors. Entries in the environment field of the PaleoDB download indicate information regarding the likely reefal origin of carbonatic rocks. See the vignette ('PhaneroCurve') on the exact use of these data. v0.9.

lith

A list of vectors. Entries in the lithology1 field of the PaleoDB download indicate information regarding the substrate of the embedding rocks. This key maps the entries to siliciclastic, "carbonate" or "unknown" substrates. v0.9.

lat

A list of vectors. Entries in the paleolat field of the PaleoDB download indicate information regarding paleolatitude of the occurrences. This key maps the entries to "tropical" or "non-tropical" latitudes. v0.9.

grain

A list of vectors. Entries in the lithology1 field of the PaleoDB download indicate information regarding the grain sizes of the depositional environment. This key maps the entries to "coarse", "fine" or "unknown" grain sizes. v0.9.

depenv

A list of vectors. Entries in the environment field of the PaleoDB download indicate information regarding the onshore-offshore nature of the depositional environment. This key maps the entries to "onshore", "offshore" or "unknown" environment. v0.9.3

Source

Stratigraphic assignments are based on the download of collection data from Fossilworks (used to be http://www.fossilworks.org/) and the dynamic time scale of the Paleobiology Database, written by J. Alroy. The assignment of numeric values were done by A. Kocsis. Environmental variables were grouped by W. Kiessling.

Details

Entries in the stratigraphic, lithological and environment fields of current Paleobiology Database downloads are too numerous to form the basis of analyses without transformations. This variable includes potential groupings of entries that represent similar characteristics. These objects can be used by the categorize function to create new variables of stratigraphic, environmental and lithological information.