The FossilWorks-based lookup table for the stratigraphic assignments of collections in the Paleobiology Database

Table including the user-chosen interval data and the stratigraphic units of the dynamic timescale.

Usage

data(stratkeys)

Format

A data.frame with 761 observations of 8 variables:

interval

The names of the registered intervals in the early_interval/max_interval and late_interval/min_interval columns.

period

The period containing the interval.

epoch

The epoch containing the interval.

X10_my_bin

The 10 million year time scale interval containing the interval.

ten

Numeric identifier of the 10 million year interval in the tens object.

stage

The stage containing the interval.

stg

Numeric identifier of the interval in the stage-level time scale provided as stages object.

Source

Used to be http://www.fossilworks.org/.

Details

Since the separation of the FossilWorks (used to be http://www.fossilworks.org/) portal from the Paleobiology Database (https://paleobiodb.org/) the access to the stratigraphic information in the database have been problematic. This table includes groupings of early_interval/max_interval entries of the dynamic timescale that users can choose during collection entry. The table assigns these intervals to some corresponding stratigraphic units from different time scales. These entries were distilled from those collections that only have a max_interval value. As there is a mismatch between the data Paleobiology Database and FossilWorks this list is not comprehensive and a couple entries are probably missing. For this reason, this dataset is expected to be updated in the future.

This particular version (v0.9.2) is based on a download of all collections in FossilWorks between the Ediacaran and the Holocene. The download took place on 22 June, 2018. The entries were transformed to keys to be used with the categorize function. Some entries were corrected manually.